Divided to teams and categories, methods moved to abstract class,watson api, updated save_binary_cache
Methods like generate output was total re-write and moved to abstract class
for better and easier access from both types of output (teams and categories)
For getting team name is used watson_api.
I was added new parameter to save_binary_cache for force write (cache are now
re-writed everytime if is this parrameter setup to True
In summariteze data was re-worked step_3 where is added check of status bug,
this is probably needed on more places in this method.

Signed-off-by: Patrik Helia <phelia@redhat.com>